Added error catching to spawn child process
If you run this package on a system that does not have a valid audio player installed and attempt to play a file, you get an uncaught error. This minor change catches that error and handles it properly allowing the project calling this module to handle the error gracefully instead of terminating.